Better splited code
