+++ src/scripts/homeScripts.ts
