+++ src/scripts/searchScripts.ts
