+++ src/scripts/yourVideosScripts.ts
